Split trimQuery() and replaceParameters() method from QueryBuilder::build()
